+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 ## [Un-Released]
 - Enhancements
     - Website use URL as the default title
+- Fixed
+    - Nonce issue after zkApp switch chain
 
 ## [2.0.1]
 - Enhancements